script refactoring

+++ bin/deinit.sh
@@ -1,6 +1,14 @@
 #!/usr/bin/env bash
 set -e
 
-git submodule deinit -f src/rexo
-rm -rf .git/modules/src/rexo
-git rm -f src/rexo
+pwd=$(pwd)
+
+install="$1"
+
+if [ "${install}" == "submodule" ]; then
+    git submodule deinit -f src/rexo
+    rm -rf .git/modules/src/rexo
+    git rm -f src/rexo
+fi
+
+cd "${pwd}"
